Improve rack-box build steps and instructions
action.yml: Reorder some steps to be consistent with instructions in
README.md.

README.md: Update instructions for downloading all files
needed to build rack-box image and make them clearer, especially how
to package the RACK documentation.

user-data: Add ppa:swi-prolog/stable to apt sources and install
swi-prolog in VM images.

clean.sh: Turn else branch into an elif branch in case
linux-cloud-tools-virtual isn't installed.  Remove commands to install
swi-prolog (they should be in user-data and install.sh, not here).

install.sh: Add ppa:swi-prolog/stable to apt sources and install
swi-prolog in Docker images.  Add comment for maintainers to keep
package lists consistent between user-data and install.sh (that is,
between VM and Docker images).

+++ rack-box/scripts/clean.sh
@@ -19,16 +19,14 @@ if [ "${PACKER_BUILDER_TYPE}" == "hyperv-iso" ]; then
     cd /usr/libexec/hypervkvpd
     ln -s /usr/sbin/hv_get_dhcp_info .
     ln -s /usr/sbin/hv_get_dns_info .
-else
+elif dpkg -l | grep -q linux-cloud-tools-virtual; then
     apt-get remove -yqq linux-cloud-tools-virtual
     apt-get autoremove -yqq # remaining linux-cloud-tools packages
 fi
 
-# Upgrade all packages, also install swi-prolog
+# Upgrade all packages
 
-add-apt-repository ppa:swi-prolog/stable
 apt-get update -yqq
-apt-get install -yqq swi-prolog
 apt-get upgrade -yqq
 
 # Clean apt cache and temporary files
